show codex patch stats in summaries
