+++ apps/renderer/src/components/turn-summary.tsx
@@ -18,7 +18,12 @@ import { groupMessages } from "../lib/group-messages.ts";
 import { cn } from "~/lib/utils";
 
 import { FileBadge } from "./file-badge.tsx";
-import { diffStats, extractEdits } from "./inline-diff.tsx";
+import {
+  diffStats,
+  extractEdits,
+  extractPatchEntries,
+  patchStats,
+} from "./inline-diff.tsx";
 import { MessageRow, type ToolResultRecord } from "./message-row.tsx";
 import { SubagentRow } from "./subagent-row.tsx";
 import { iconForTool } from "./tool-row.tsx";
@@ -39,19 +44,32 @@ interface FileStat {
 
 const aggregateFileStats = (body: ReadonlyArray<Message>): FileStat[] => {
   const map = new Map<string, { added: number; removed: number }>();
-  for (const m of body) {
-    if (m.content._tag !== "tool_use") continue;
-    const tool = m.content.tool;
-    if (tool !== "Edit" && tool !== "Write" && tool !== "MultiEdit") continue;
-    const edits = extractEdits(tool, m.content.input);
-    if (edits.length === 0) continue;
-    const stats = diffStats(edits);
-    const path = edits[0]!.path;
+  const addStats = (
+    path: string,
+    stats: { added: number; removed: number },
+  ) => {
     const prev = map.get(path) ?? { added: 0, removed: 0 };
     map.set(path, {
       added: prev.added + stats.added,
       removed: prev.removed + stats.removed,
     });
+  };
+  for (const m of body) {
+    if (m.content._tag !== "tool_use") continue;
+    const tool = m.content.tool;
+    if (tool !== "Edit" && tool !== "Write" && tool !== "MultiEdit") continue;
+    const patches = extractPatchEntries(m.content.input);
+    if (patches.length > 0) {
+      for (const patch of patches) {
+        addStats(patch.file_path, patchStats([patch]));
+      }
+      continue;
+    }
+    const edits = extractEdits(tool, m.content.input);
+    if (edits.length === 0) continue;
+    const stats = diffStats(edits);
+    const path = edits[0]!.path;
+    addStats(path, stats);
   }
   return Array.from(map.entries()).map(([path, s]) => ({ path, ...s }));
 };
